Implement computation primitive in standard kernel
 * Add Compute.{sig,sml} to src/0/, containing an ML interpreter for
   first-order lisp-style pairs, and code for translating from normal
   HOL terms into this datatype and back.
 * Add wrapper code to the Thm signature to call the compute primitive
 * Add new counters for the Compute rule to src/prekernel/Count.{sig,sml}
 * Update the theorem signature

This commit contributes a HOL4 port of Candle's compute primitive. The
compute primitive is an interpreter for first-order lisp-style pairs of
numbers with arithmetic.

Candle's compute primitive has been verified sound. The original Candle
implementation and its mechanized verification proof can be found at:

  https://code.cakeml.org/tree/master/candle/prover/compute

This port attempts to stay as faithful as possible, but some things are
a bit different in HOL4:

  * HOL numerals use a slightly different representation from
    Candle/HOL Light
  * Division by zero is defined in Candle/HOL Light
  * In Candle, the characteristic equations used by the compute
    primitive are defined using HOL4 overloads. We solve this using
    recognizers that resemble the overloads as closely as possible.

The contributed compute primitive needs to create theorems about
arithmetic and datatypes, but such things are derived long after the
kernel is loaded. To solve this, the primitive needs to be instantiated
with a record of terms, types and theorems at runtime. The terms and
theorems correspond to the syntax on which the compute primitive will
operate, and the theorems are axioms about these symbols. For example,
the primitive needs to know about the type :num, the symbols SUC and
(+), and that SUC m + n = SUC (m + n).

The compute primitive adds roughly 700 loc to the standard kernel. Most
of this code is about deconstructing and reconstructing HOL4 terms
(e.g., dest_numeral/mk_numeral), and recognizing various terms (e.g.,
SUC, NUMERAL, +, ...) as these things are not available when working in
the kernel. The actual interpreter used to reduce expressions and its
datatype is around 50 loc.