Add syntax, code and examples using Thm.compute
* Define the datatype :cv of pairs of numbers used by the kernel
  compute primitive, and all functions recognized by Thm.compute
* Add wrapper library (cv_computeLib) that initializes Thm.compute
* Add some usage examples

This commit contributes all the theories and code required to initialize
and use the kernel primitive (Thm.compute), as well as some examples of
basic usage, and a selftest.

The wrapper is called cv_computeLib and provides an initialized version
of Thm.compute, with this type signature:

  cv_compute : thm list -> term -> thm

where the thm list is a list of code equations, and the term is the term
to be reduced. See examples/examplesScript.sml for an explanation.