Make tcall and trec constants dramatically simpler
Now tcall is of type

  : (α -> β + γ) -> (β -> γ) -> (α -> γ)

The first argument is the function that acts as a combined guard and
computation of either the "next state" (left), or the terminating
state (right). The second function (the thing that is tail-called)
takes the "next state" value and maps into the desired final type.

Then, on top of this trec can then be of what is clearly the "right"
type:

  : (α -> α + β) -> α -> β

so that trec_thm says:

  trec c x = tcall c (trec c) x

(read "c" as "code").