Forbid "rebinds" of exported theorems
This is not enforced interactively of course, when users will be
replaying definitions and proofs constantly. The [allow_rebind]
annotation can be used when the rebind is actually wanted, but as all
the fixups in this commit demonstrate, this change catches a great
number of errors.

Resolving all the errors is still ongoing.