Make thm-binding after {Co,}Inductive def'ns quieter if erroring
If the basic inductive definition goes wrong, subsequent calls to cj
on the rules theorem (performed to extract individual rules) may well
fail too, but hearing all about these failures is an unhelpful
distraction when the real problem is elsewhere. This change wraps the
calls to cj with a handler that makes the theorems just be TRUTH in
the event of failure.

Thanks to Magnus Myreen for the suggestion.