HolSmt: reenable some 'real' tests for Z3 proof replay
According to a comment in the list of self-tests, Z3 v2.19 prints real numerals
in the same format as integers in its proofs, which means it may not be possible
to distinguish between them. This also means it may not be possible to replay
most Z3 proofs containing reals.

This issue seems to have been fixed in Z3 v4.12.4 (and likely, many earlier
versions as well) but we don't support proof reconstruction for modern Z3
versions yet.

Despite this issue, some tests do seem to replay correctly with Z3 v2.19, which
is the only version of Z3 that we support for proof replay/reconstruction.

Since for the moment we can assume that Z3 v2.19 is being used for Z3_TAC (as
opposed to Z3_ORACLE_TAC), let's reenable these tests to avoid regressions.