HolSmt: add support for the reals' abs, min and max
Hopefully this is a temporary stop-gap measure until a more general mechanism
is implemented to identify an appropriate definition that can be sent to the
SMT solvers, rather than special-casing each and every definition.

That said, this is still useful, especially so that we can enable some of the
already-existing self-tests for the SMT-LIB translation as well, to detect
future regressions whenever improvements are made.

The enabled self-tests all include Z3 with proof reconstruction now.