HolSmt: add some support for tuples
Again, this is a temporary stop-gap measure until we can translate function
definitions to SMT-LIB format, so that SMT solvers can reason about them.

As it is, due to the above limitation this won't be able to solve some goals
(witness the two disabled tests) but it does allow solving the simpler ones.

Of the tests that were enabled, all work with Z3 proof reconstruction.