HolSmt: add support for parsing the Z3 v4 proof grammar
Support for old-style Z3 v2.19 proofs is retained, since the
differences are minor.

That said, in the future it might make sense to remove support for Z3
v2.19 proofs due to such an old version being obsolete, although it'd
be desirable for that to happen only after Z3 v4 proof replay reaches
feature parity.