HolSmt: handle the intro-def rule in Z3 v4 proof replay
This rule basically introduces a name for a term, i.e. a definition
of the form:

``var = term``

However, Z3 often defines such a variable by providing a term in a
more complicated form such as (for example):

``(var \/ ~term) /\ (~var \/ term)``

(All such schematic forms provided by Z3 are tautologies given that
`var = term`).

We track all such definitions, which accumulate in the set of
hypotheses of all theorems that depend on them, and at the end they
are all removed from the final theorem.

We can remove them from the final theorem because these variables do
not appear in the remaining hypotheses nor in the conclusion of the
theorem, i.e. all such variable names are invented by Z3 for the
purposes of facilitating intermediate proofs and are thus irrelevant
with regards to the final theorem.