HolSmt: unify terms in the rewrite rule handler
While reconstructing a proof of `(if x < y then x else y) <= x`,
Z3 v4.12.4 asks us to prove the following rewrite rule as one of the
proof steps:

(let ((@x91 (rewrite (= (not (>= (+ x (* (- 1) (ite $x36 y x))) 0))
  $x86))))

... corresponding to this term:

~(x + -1 * (if x + -1 * y >= 0 then y else x) >= 0) <=>
~(x + -1 * $var$(z3name!0) >= 0)

... where z3name!0 is a variable declared by Z3 at the beginning of
its proof certificate, but which we know nothing about at this point.

To make the proof go through, we unify the two sides of the
equivalence, such that we obtain:

z3name!0 = (if x + -1 * y >= 0 then y else x)

We then prove the theorem by substituting the variable(s) in the goal
and add the above definition to the same list of Z3-provided
definitions as that used by the `z3_intro_def` handler, so that it
gets removed at the end of the proof.