HolSmt: fix bitvector operator parsing in Z3 v4 proofs
This commit fixes the parsing of the `bvand`, `bvor`, `bvadd`,
`bvmul`, `bvxor` and `bvxnor` operators in Z3 v4 proof reconstruction.

`bvxor` and `bvxnor` were already implemented but only as binary
operators.

However, Z3 v4.12.4 declares `bvxor` and `bvxnor` as being "flat
associative" and in (at least) one of the Z3 proofs in the self-tests
`bvxor` does appear with 3 arguments. Therefore, the parser was
changed to parse these operators using a left-associative application
of the arguments.

The other operators were missing, i.e. they were not being parsed.
In z3 v4.12.4 they are similarly declared as being flat associative.

Strangely, this version of Z3 does not declare `bvnand` and `bvnor` as
being flat associative, so we keep parsing them as binary operators.