HolSmt: deprecate Z3 v2.19 in favor of v4.12.4
Now that proof reconstruction for Z3 v4.12.4 works, we can deprecate
Z3 v2.19.

Using a modern version of Z3 has many advantages, since Z3 is now
open-source and also, binaries exist (or can be compiled) for
platforms other than x86 and x86-64.

Also, Z3 v4.12.x is in active development, which means there's a
greater chance that bugs that are encountered will be fixed (unlike
Z3 v2.19 which is a very old, deprecated version) and other
improvements made.

Modern versions of Z3 are also easier to install in current Linux
distributions, since usually they are already part of their package
collection.

Furthermore, if we can ignore Z3 v2.19 then we can start enabling
proof reconstruction for those self-tests which can only be
reconstructed in Z3 v4 (once the reconstruction is fixed, that is).