HolSmt: new implementation of `remove_definitions`
This new implementation was developed to address concerns about
exponential term blow-up.

Basically, if we have definitions of the form:

z1 = 1
z2 = z1 + z1
z3 = z2 + z2
...
z64 = z63 + z63

... then the old implementation would have fully expanded the
definition of `z64` in order to remove it, which would not fit in
memory due to the 2^64 variable substitutions which would be
required.

This new implementation avoids fully expanding definitions. It was
developed with valuable suggestions and support from @tjark.