HolSmt: add support for the `num` type
This commit adds support for SMT solvers to reason about the `num`
type by providing them with some theorems which allow them to convert
`num` operations into integer operations which they natively support.
Specifically, it adds to the proving context (as assumptions) theorems
for the following conversions between `num` terms and integers:

* `num` literals [INT_OF_NUM]
* `SUC`, i.e. ``&SUC n <=> &n + 1i`` [INT]
* `num` equality [INT_INJ]
* `num` inequalities [INT_LE, INT_LT, GREATER_DEF, GREATER_EQ]
* Non-negativity of `num` terms [INT_POS]
* `num` addition and multiplication [INT_ADD, INT_MUL]
* MIN and MAX [INT_MIN, INT_MAX]

This is not an exhaustive list of all the theorems that will be
needed to reason about the `num` type or even just `num` arithmetic,
but it does allow Z3 to solve all the `num`-related self-tests, with
the notable exception of DIV and MOD operators. To fix the latter, we
will need to add support for integer division and mod first.

Unfortunately cvc5 cannot solve a few of the `num` tests and the Z3
proof reconstruction fails for almost all `num` tests, although this
is not due to unsoundness in `num` support, but rather because the Z3
proof certificates for these goals are more complex and use new proof
syntax and rules which we do not support yet (this will be fixed
later).

There is another unfortunate regression: adding these theorems to the
context prevents SMT solvers from giving `sat` answers. In fact, this
happens even when just one theorem is added to the goal as an
assumption (`integerTheory.INT`, specifically, which is needed to
reason about `SUC`), regardless of which goal we try to prove.

Fortunately, this does not prevent them from answering `unsat`.

Since in the self-tests we do rely on `sat` answers to validate
unprovable goals, the tests which expect a `sat` answer were modified
to temporarily disable the inclusion of the above theorems while
performing those tests.