HolSmt: disable Z3 v2 on failing test cases
Z3 v2 is now failing in some tests due to fallout of some previous
changes that were needed to add support for the `num` type.

There are 2 root causes for these failures:

1. Z3 v2 doesn't know about the integer `ABS` function. We used to
rewrite the conclusion of the goal with the `INT_ABS` theorem but this
is not needed when using modern SMT solvers because SMT-LIB supports
`ABS` natively.

2. Due to changes in how some goals are passed onto the SMT solvers,
Z3 v2 is now generating proof certificates with some undeclared and
untyped terms which we are not able to reconstruct.

While I don't yet see a compelling reason to actively remove support
for Z3 v2 (since support for Z3 v4 is mostly a superset of our support
for Z3 v2), I don't see a compelling reason for adding v2-specific
workarounds either, due to how old it is and due to it being
unsupported for many years already.

This commit adds new `thm_Z3_v4` and `thm_Z3p_v4` functions which only
run the test cases if Z3 is configured and is version 4. These new
functions are used in the test cases which succeed with Z3 v4 but not
with Z3 v2.