HolSmt: parametrize the SMT-LIB parser by `let` handling functions
The SMT-LIB term parser and Z3 proof certificate parser are similar
in most ways, but they need to handle `let` terms differently.

To avoid having to duplicate a significant portion of the parser,
this commit parametrizes the SMT-LIB parser so that it can handle
these `let` expressions differently based on parser handling
functions that are passed by the caller of the term parser.

In this commit, only the SMT-LIB version of the `let` handling
functions are implemented, so it doesn't change how the parsers work
yet.

However, in a subsequent commit, the Z3 proof parser will use this new
functionality to parse the proof certificates correctly.