HolSmt: parse `lambda` binders in Z3 proofs
These seem to be a local declaration of variable names+types that
apparently are meant to be interpreted as free variables in the
enclosed term.

Note that these `lambda` binders should only be parsed when parsing
Z3 proof certificates, as the SMT-LIB language as specified in the
current 2.6 version of the standard doesn't actually have these
binders and we don't want to parse identifiers coincidentally named
`lambda` as these special binders in this case.

We take advantage of the previously-implemented parser parametrization
to determine whether `lambda` binders should be parsed or not.