HolSmt: fix `quant-intro` proof rule to work w/ `exists` quantifiers
The `quant-intro` rule handler could only handle `forall` quantifiers.
However, Z3 also uses it for `exists` quantifiers.

This commit makes the handler work with both quantifiers, including
when they are interspersed with each other.

This fix is necessary for replaying a `quant-intro` rule in the Z3
proof for the following self-test:

?x. x = x