HolSmt: fix `elim-unused` proof rule to work w/ `exists` quantifiers
The `elim-unused` rule handler could only handle `forall` quantifiers.
However, Z3 also uses it for `exists` quantifiers.

This commit makes the handler work with both quantifiers, including
when they are interspersed with each other.

This fix is necessary for replaying an `elim-unused` rule in the Z3
proof for the following self-test:

?x. x = x

This commit also enables proof replay for the above self-test, since
the proof can now be fully reconstructed.