HolSmt: fix Z3 proof replay of `asserted` proof rules
This commit contains both a fix and a workaround for replaying
`asserted` proof rules.

The fix consists in disabling the `pp.simplify_implies` Z3 parameter.
This parameter controls whether the Z3 AST pretty-printer does some
simplifications when printing the proof and is enabled in Z3 by
default.

However, these simplifications can effectively mangle the terms in the
`(asserted ...)` inference rules, which means they will not match the
terms that were actually asserted in the input file (i.e. the terms in
the assumptions of the goal). Therefore, these mangled terms would
cause proof replay to fail in some cases.

Unfortunately, in currently released versions of Z3 (i.e. 4.12.6 and
earlier) it is not actually possible to disable `pp.simplify_implies`
when printing the proof. A fix for this issue was submitted as the
following PR:

https://github.com/Z3Prover/z3/pull/7157

To mitigate this parameter not currently working in Z3, a workaround
was also implemented which tries to remove mangled asserted
statements, which would end up as extra hypotheses in the final
theorem.

This workaround has been confirmed to work in Z3 v4.12.6 as well as Z3
v2.19.

Since this issue has been fixed, we can now enable two more self-tests.