HolSmt: add some arithmetic theorems
SMT_NUM_ADD and SMT_NUM_SUB were meant to be used in place of INT_ADD
and INT_NUM_SUB for optimization reasons.

However, this causes a regression in proof replay in one of the
self-tests. Since this regression is not straightforward to fix,
these theorems are not being used yet.

NUM_OF_INT was also added to the default set of theorems since it
helps solve some `num`-related goals.