HolSmt: fix workaround for extra hypothesis in proof replay
The previously-implemented workaround might have an issue where if Z3
introduces an hypothesis of the form `P = P` but then *doesn't* forget
to discharge it, we might fail to replay the `lemma` proof rule
because we avoided adding an assumption.

Instead, this commit fixes the workaround so that we remove the extra
hypothesis at the end of the proof replay procedure, i.e. only when we
are sure that Z3 *did* forget to discharge the hypothesis.