Fix bug in Omega test
Make slight changes to code along the way; making it easier to see
state of dictionary that is used internally.  Bug was in call to
one_var_analysis where a stale map variable (ptree) was used, when the
correct parameter was (pt) which was the continuation's parameter.

Add a number of test-cases, all of which are probably more or less
equivalent.

Thanks to someplaceguy for the bug report!

Closes #1209