HolSmt: replay goals containing `quot` and `rem`
HOL4's integer arithmetic decision procedures cannot handle `quot` and
`rem`, so proof steps containing these symbols were failing to replay.

This commit rewrites goals containing these symbols to increase the
chances of a successful replay.

`thm_AUTO` was similarly improved, as it had the same limitation and
was failing to prove some of the self-tests.