HolSmt: don't get rid of `quot` and `rem` in arithmetic rules
It should not be necessary to rewrite these symbols with their
definitions, because from the perspective of SMT solvers, these should
just be functions like any others. Therefore, when replaying
arithmetic proof steps, the proof handler should not need to
understand what `quot` and `rem` are.

This is a partial revert of the following commit:
6220a17848fd58bcb54bb9df9df6b55d5859abf0

It's not clear to me why I introduced this change in the first place,
since the tests do work fine without it. Also, there doesn't seem to
be any performance regression.