HolSmt: allow SMT solvers to peek under abbreviations
Otherwise, they won't be able to see what the variables are being abbreviated
to. This problem would cause some goals that can be solved with metis_tac[] to
fail with HolSmt.

The fix is to add the definition of the `Abbrev` symbol to the list of theorems
that are added to the assumptions prior to calling the SMT solver.