HolSmt: add z3_tac and z3o_tac tactics
These accept a list of theorems, like `metis_tac`. The free vars in the theorems
are automatically quantified with `Drule.GEN_ALL` and added to the list of
assumptions in the goal, prior to sending the goal to the SMT solver.

This is needed to avoid any passed-in theorems to be specialized to free vars in
the goal that happen to have the same name, which is almost always not what the
user wants (and if the user really wants it, they can always assume such
theorems manually if needed). This matches the behavior of `metis_tac`, thus
avoiding unwanted surprises.