HolSmt: initial support for the exponential function
This initial support is very limited and only consists in adding a
couple of useful theorems to the proving context.

Hopefully, in the future more extensive support is implemented by
adding support for Z3's native power operator, which should allow
HolSmt to prove a lot more goals. This implementation would need to be
similar to the implementation of integer division, since it doesn't
exactly match HOL4's exponential functions.

The latter would also have a couple of limitations, namely:

1. it would only work for Z3, since this operator is a Z3 extension
and not part of the SMT-LIB standard
2. proof reconstruction would not work