HolSmt: use new constant for SMT-LIB's real division
The `realax$/` constant is probably not safe to be used directly as
the SMT-LIB real division operator because it defines `x / 0r` to be
zero, which (as I understand) might not necessarily be true in the
SMT-LIB's `Reals` theory.

In the case of Z3, it seems to consider `x / 0r` undefined, which does
not present a soundness issue, but my interpretation is that SMT-LIB
allows implementations to define the result of this operation -- and
if they do, strictly speaking there's no guarantee that it will be
defined to be zero.

Consequently, a new `smt_rdiv` constant was defined to properly model
SMT-LIB's version of real division in HOL4, which leaves the result of
`x / 0r` undefined.

A theorem was also added to the proving context so that SMT solvers
can convert `realax$/` operations into `HolSmt$smt_rdiv` operations
(and back).

As a result of this, it should now be possible for SMT solvers to
prove `x / 0r = 0r`, which is true for all `x` in HOL4.