HolSmt: add theorems selectively
This commit makes HolSmt add theorems to the proving context
selectively.

That is, instead of always adding all theorems that might be needed,
HolSmt will now look at all the terms in the goal and only add those
theorems that are necessary to solve this particular goal.

As an example, for a goal such as `1i + 2 = 3`, HolSmt used ASSUME_TAC
to add a bunch of unrelated theorems to the proving context, such as
realaxTheory.real_pow, realTheory.abs, etc. These were not actually
needed to solve this particular goal, although they are needed to
solve other goals.

After this commit, HolSmt no longer adds all these unnecessary
theorems. This allows SMT solvers to solve goals more efficiently, in
some cases.

To select which theorems to add, HolSmt now looks at all the constants
used in all the terms in the goal, and then, for each constant, it
adds the relevant theorems. Furthermore, it iteratively does the same
process for all theorems that were just added, since they themselves
might contain constants that require other theorems to be added.