transferLib enhancements
Runs faster, copes with more examples, and has new handling of terms
that aren't supposed to be examined internally (like numerals). These
are called "atomic" and can be registered by the user. This is used
in the pure de Bruijn term example to mask out the nominal complexity
lurking underneath the FV "constant".