HolSmt: fix parsing of Z3 proof certificates
Apparently, some Z3 proof certificates contain indexed identifiers
which are not valid SMT-LIB terms (even in new Z3 versions), such
as `-3` and `1/2`.

This commit adds a new parsing function that successfully parses
these identifiers (as reals, although they are currently ignored for
the purposes of proof reconstruction).

A new `th-lemma arith` identifier called `gomory-cut` (which is also
ignored for the purposes of proof reconstruction) was also added so
that the parser doesn't error out.