HolSmt: don't use `--finite-model-find` with cvc5
This option was introduced in the following commit, to avoid
regressions in the self-tests after introducing the ability to pass
relevant theorems to the SMT solvers:
7f0fff50f078f56baf7787ea58a1a64db0833408

However, currently this option seems to be hurting cvc5's performance
more than it is helping. Therefore, this commit removes it.

As a result, there are 2 regressions in the self-tests, but on the
bright side, many dozens of disabled self-tests can now be enabled for
cvc5.