Make `buildheap --exe` not create a temporary `.ML` file
In some situations, these temporary `.ML` files would be left laying
around the HOL source tree instead of being deleted, even when
buildheap's `--dbg` option is not being used.

This could happen for multiple reasons, namely: due to a system crash
or abrupt power off, the user or the OS killing the buildheap process,
or more commonly, due to Holmake killing buildheap because some other
build step that was running in parallel failed.

The result of the above situations is that these `.ML` file(s), which
are untracked by git, suddenly appeared in the HOL source tree. More
importantly, the user might not notice this happening until much
later, at which point he might no longer realize what was the cause of
the problem or whether they can be deleted.

This commit removes the creation of the temporary `.ML` files
altogether, calling the Poly/ML compiler directly instead to create
the compiled object file.

Note that this commit does not remove the similar creation of the
temporary `.o` object files, created by `PolyML.export(...)`.

Initially, it was considered that `OS.FileSys.tmpName ()` could be
used to create these object files, but this option was rejected for
the following reasons:

1. `tmpName ()` does not allow the calling code to choose the file
   extension, and indeed, in Poly/ML the resulting files don't have
   an extension.

2. It does not seem possible to extend or recreate `tmpName ()`
   securely within SML, in a way that allows choosing an extension
   (although, it might be possible to do it with the FFI, but the
   effort does not seem worthwhile).

3. According to the documentation, the `cc` command typically expects
   object files to have a `.o` extension, to differentiate them from C
   source files. In practice, both gcc and clang's `cc` command seem
   to link object files successfully even when they don't have a `.o`
   extension, but this seems to be undocumented (thus probably not a
   good idea to rely on this behavior) and it might not work for other
   compilers.

Fortunately, leftover `.o` files are not as bothersome as leftover
`.ML` files, because the former have an entry in `.gitignore`, thus
being hidden from most git commands.