Add machinery to let Q.RENAME_TAC (rename) specify match style
Also fix up a number of uses of

  Q.MATCH_{GOAL,ASM}SUB_{RENAME,ABBREV}_TAC

to use the versions that attack whole goals or assumptions rather than
bother to look for sub-terms. At an earlier stage of development of
this work, I had to do this because of a behaviour change that has
since disappeared, but it's still better, if possible, to use

  Q.MATCH_ABBREV_TAC or Q.MATCH_ASSUM_RENAME_TAC or ...

because of the additional precision (and these names are shorter to
boot).

Some documentation and tests.