Reinstate old behaviour when rewrites compete
If we have ⊢ x = e1 in the argument list and x = e2 as an assumption,
the theorem in the assumptions used to be chosen first by simp and
friends. The commit in cfa24e268b flipped this choice, but this flips
it back again. Of course, users shouldn't really be relying on this,
but at least one CakeML proof did, and in this case, there seems no
need to punish this.