Move LT_EXISTS and LE_EXISTS theorems to arithmetic only
Previously, there were variations of these scattered around src
unnecessarily.

There are some minor knock-on changes, but the bulk of the changes in
this commit are me trying to make proofs more robust in the face of
possibly changing LE_EXISTS's existential variable's name from p to
d (for "delta", and matching what is done in LT_EXISTS).

This broke a few things, most annoyingly when the `e + d` or `e + p`
was rewritten by the standard bossLib simplifier. Because this
includes the arithmetic normalisation code, the summands get sorted,
and d sorts to different positions than p.

So, I ended up keeping the old name (p), but I'm happy to also keep
the robustifications included in this commit.