fix for DataSize error for finite_map
[github issue #1400]

A proof of type-size equivalence fails when the finite_map type is present,
because the size installed for finite_map contains a lambda and causes a
BETA_TAC to apply to the wrong thing. This simple fix seems to solve the
problem.

The breaking example depends on the finite_map type, and probably can't be
easily added to the self-test.