Extend HolParser's #(LINE=) directive to optionally include column #
This can then be used by the emacs mode to set buffer-relevant
positions for stuff sent to an interactive session. This is necessary
because the switch to HolParser removed the old use of
"loc-pragmas" (comments of form (*#loc line col *)) to do the same.

Unfortunately, column numbers are off (and were off before) because
they count raw characters rather than UTF8 code points.