Make repl test Holmakefile specify hol.state0 and be only Poly/ML
This test is only relevant for Poly/ML HOL, so (almost) the whole file
may as well be ifdef POLY ... endif bracketed. Actually keep
EXTRA_CLEANS visible so that a subsequent mosml build/Holmake can at
least clean the directory.

Specifying hol.state0 is not necessary for build, which arranges that
anyway, but interactively, it's useful to be able to call Holmake in
this directory without needing to give the --holstate=.. option
manually.