Fix 1563 (#1568)
* Fix #1563

Seems the main culprit was a minor mistake in the lexing rules.
After Ancestors or Libs we expect attributes to follow, not
an identifier with attributes.

The change to holstartkeyword feels a bit hacky; additionally,
without those parentheses around the added patterns things seem
to break for some reason.

* Use Theory syntax in stringScript

Hopefully this makes regressions easier to catch in the future.

* [HolLex] Use attributes in id_with_attributes

* [HolLex] Move attributes out of holstartkeyword

holstartkeyword is intended to be just the keyword list and may be used in other places in the future -- we deal with the conflict in headerthy instead, where it actually arises.

* [stringScript] Drop ignore_grammar