Use Theory in src/bag
Some files have set_grammar_ancestry, but do not mention all of
their ancestors. This could be because it is unnecessary (since
some ancestors may already be part of the grammar thanks to some
other ancestors), or it could be to explicitly not include it.

For now, we will assume that it's okay to add ancestors to the
grammar. However, for debugging purposes, I will try to maintain
a list of files whose grammar may (or may not) include more now.

Files whose grammar maybe include more now:
- src/bag/bagScript.sml