[HolLex] Allow qualified identifiers in Theory header
This is necessary for bool, since it depends on Unicode, which
is a substructure of Parse since commit 721dd13.
In particular, we want to be able mention Parse.Unicode, since
by the time we hit Unicode, I assume that the relevant bindings from
Parse are not yet available.

Thanks to Irvin (@ordinarymath) for finding the relevant commit
and explaining how splitting open could make a difference.