Add holTheory and use it in Theory syntax
To my understanding, a normal hol session has at least indexedLists and
patternMatches as parents:
```
> parents "-";
val it = ["indexedLists", "patternMatches"]: string list
```

It also appears that some theories depend on this. For example, if one were to
take countable_types from pred_set, and set its grammar ancestry to the theories
it explicitly opens (bool and pred_set), the proofs will break. This is
problematic, as this is exactly what the Theory syntax does.

This PR/commit deals with this issue by
- creating a new theory called "hol", which represents the standard context of
something starting in the hol heap
- rename `hol` to `lambda_hol` in `examples/lambda` to resolve the name conflict
- changes the Theory syntax to load/open holTheory, and include hol in
set_grammar_ancestry calls