Add LAST_ASSUME_TAC and adjust ASSUME_NAMED_TAC (#1580)
* Add LAST_ASSUME_TAC (and last_assume_tac)

* LAST_ASSUME_TAC in ASSUME_NAMED_TAC

* make ASSUME_NAMED_TAC smarter

The idea is that named assumptions should be at the top
but the addition of new named assumptions should not be higher
than existing named assumptions.
This two examples show the difference.
ASSUME_NAMED_TAC "hi" TRUTH ([``bar``],``foo``)
ASSUME_NAMED_TAC "hi" TRUTH ([``baz :- T``],``foo``)
Note that this code is designed to ignore when there's a named example in the middle
eg. in this case bar is ignored
ASSUME_NAMED_TAC "hi" TRUTH ([``foo``,``bar :- T``,``baz``],``foo``)
Note that this is implementation is fairly inefficient for code with huge number of assumptions
but should not be encountered in practice.

* Update release notes

---------

Co-authored-by: ordinarymath <ordinarymath@yahoo.com>