Mention minor new incompatibility in Moscow ML help system
See #1600 to see me realising why this is impossible to do, given the
simple-minded implementation of hol as `unquote | mosml ...`.

In the Poly/ML filter, the quotation filter machinery applies (only)
to text that is being fed to the compiler; if the compiled and
executed code then reads from stdin, it will get those characters
unfiltered.  The Moscow ML version filters *everything*.