[emacs-mode] "quiet" region send avoids uninteractive export_theory
When toggling "quietness", the code sets Globals.interactive to be
false.  Unfortunately this caused theories to get exported if one has
selected all of the prefix of a modern theory with its "Theory"
header. Now, the interactive flag is set back to true just before the
export_theory call gets made (instead of after), and so this ensures
that the export doesn't happen.