Get theorytool working again given shift to .hol (not .holobjs)
The "tool" doesn't do much at present, but its loading and hashing of
theory files now runs "fast enough" (400 theories in < 2s) so I don't
think caching hash results is necessary.

This tooling should now be capable of generating the theory graph, but
the existing graph generation not only generates the graph but also a
pile of HTML presentations of theories.